refactor: lib 디렉토리 구조 재정비 (firebase 분리, server/admin/shared 역할 구분)
