+++ components/auth/AdminProtectedRoute.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { useRouter } from "next/router";
-import { clearAuthCookie } from "@/lib/cookies";
+import { clearAuthCookie } from "@/lib/server/cookies";
 import axios from "axios";
 import LoadingSpinner from "@/components/common/LoadingSpinner";
