+++ hooks/useFetchRecommendations.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { collection, getDocs, query, orderBy } from "firebase/firestore";
-import { db } from "@/lib/firebase";
+import { db } from "@/lib/firebase/client";
 
 export function useFetchRecommendations() {
   const [recommendations, setRecommendations] = useState([]);
