+++ hooks/useRecommendationDetail.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { doc, getDoc } from "firebase/firestore";
-import { db } from "@/lib/firebase";
+import { db } from "@/lib/firebase/client";
 
 export function useRecommendationDetail(id) {
   const [data, setData] = useState(null);
