+++ hooks/useUploadImage.js
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { uploadImage } from "@/lib/uploadImage";
+import { uploadImage } from "@/lib/firebase/uploadImage";
 
 export function useUploadImage() {
   const [uploading, setUploading] = useState(false);
