+++ lib/firebase/client.js
