+++ lib/firebase/uploadImage.js
@@ -1,5 +1,5 @@
 import { ref, uploadBytes, getDownloadURL } from "firebase/storage";
-import { storage } from "@/lib/firebase";
+import { storage } from "@/lib/firebase/client";
 import { v4 as uuid } from "uuid";
 
 export async function uploadImage(file, folder = "recommendations") {
