+++ lib/server/auth.js
