+++ lib/server/deleteUserData.js
@@ -1,4 +1,4 @@
-import { db, adminAuth } from "@/lib/firebaseAdmin";
+import { db, adminAuth } from "@/lib/firebase/admin";
 
 /**
  * 관리자 또는 사용자 탈퇴 시, 계정과 관련된 모든 데이터 삭제
