+++ lib/server/users.js
@@ -9,7 +9,7 @@ import {
   query,
   Timestamp,
 } from "firebase/firestore";
-import { db } from "@/lib/firebase";
+import { db } from "@/lib/firebase/client";
 
 /**
  * users_public + users_private 병합된 사용자 목록 조회
