+++ lib/shared/utils.js
