+++ pages/admin/reports/chats.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { onAuthStateChanged } from "firebase/auth";
-import { auth } from "@/lib/firebase";
+import { auth } from "@/lib/firebase/client";
 import LoadingSpinner from "@/components/common/LoadingSpinner";
 import ChatReportTable from "@/components/reports/tables/ChatReportTable";
 import ChatReportDetail from "@/components/reports/details/ChatReportDetail";
