+++ pages/api/admin/report/chat/deleteWithReports.js
@@ -1,5 +1,5 @@
-import { db } from "@/lib/firebaseAdmin";
-import { verifyAdminToken } from "@/lib/auth";
+import { db } from "@/lib/firebase/admin";
+import { verifyAdminToken } from "@/lib/server/auth";
 import admin from "firebase-admin";
 
 export default async function deleteMessageWithReports(req, res) {
